Move jujud/run to internal/run
As the run package depends on the cmd package, keep it in the
cmd package under the internal package. This should prevent people
importing it outside of the cmd package.
